separate bottom upload
