+++ src/components/molecules/index.js
@@ -6,9 +6,11 @@ import DefaultAds from './DefaultAds';
 import CardCategory from './CardCategory';
 import Upload from './UploadFoto';
 import Menu from './MenuAkun'
+import BottomUpload from './BottomSheetUpload';
 
 export {
   Upload,
+  BottomUpload,
   Menu,
   ButtonNavigatior,
   CardSeller,
