+++ src/screens/InfoAkun/styles.js
@@ -7,31 +7,4 @@ export default StyleSheet.create({
     marginTop: 24,
     marginBottom: 30,
   },
-  bSheet: {
-    padding: 20,
-    backgroundColor: COLORS.white,
-    paddingTop: 20,
-  },
-  bSheetContainer: {
-    backgroundColor: COLORS.white,
-    paddingTop: 20,
-    borderTopWidth: 1,
-    borderLeftWidth: 1,
-    borderRightWidth: 1,
-    borderTopLeftRadius: 20,
-    borderTopRightRadius: 20,
-    borderColor: COLORS.gray1,
-  },
-  bSheetHeader: {
-    alignItems: 'center',
-  },
-  bSheetTitle: {
-    ...FONTS.h2,
-    height: 35,
-  },
-  bSheetSubtitle: {
-    ...FONTS.h4,
-    height: 30,
-    marginBottom: 10,
-  },
 });
